# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gedilson Rai Barbershop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: string;
  name: string;
  price: number;
  priceLabel: string;
  durationMinutes: number;
  durationLabel: string;
  active: boolean;
  sortOrder: number;
}

export interface CreateServiceBody {
  /** URL-friendly unique ID (e.g. corte-simples) */
  id: string;
  name: string;
  price: number;
  durationMinutes: number;
  sortOrder?: number;
}

export interface UpdateServiceBody {
  name?: string;
  price?: number;
  durationMinutes?: number;
  active?: boolean;
  sortOrder?: number;
}

export interface BlockedSlot {
  id: number;
  date: string;
  time?: string | null;
  reason?: string | null;
  allDay: boolean;
  createdAt: string;
}

export interface CreateBlockedSlotBody {
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Specific time slot to block (HH:MM). If null, blocks entire day. */
  time?: string | null;
  reason?: string | null;
  allDay?: boolean;
}

export interface Barber {
  id: number;
  name: string;
  photo?: string | null;
  phone?: string | null;
  birthDate?: string | null;
  bio?: string | null;
  specialty?: string | null;
  instagram?: string | null;
  active: boolean;
  createdAt: string;
}

export interface CreateBarberBody {
  name: string;
  photo?: string | null;
  phone?: string | null;
  specialty?: string | null;
}

export interface UpdateBarberBody {
  name?: string;
  photo?: string | null;
  phone?: string | null;
  birthDate?: string | null;
  bio?: string | null;
  specialty?: string | null;
  instagram?: string | null;
  active?: boolean;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Appointment {
  id: number;
  clientName: string;
  clientPhone: string;
  serviceId: string;
  serviceName: string;
  servicePrice: number;
  date: string;
  time: string;
  status: AppointmentStatus;
  barberId?: string | null;
  isRecurring: boolean;
  recurrenceType?: string | null;
  recurrenceGroupId?: string | null;
  createdAt: string;
}

export interface CreateAppointmentBody {
  clientName: string;
  clientPhone: string;
  serviceId: string;
  date: string;
  time: string;
  barberId?: string | null;
}

export type UpdateAppointmentBodyStatus =
  (typeof UpdateAppointmentBodyStatus)[keyof typeof UpdateAppointmentBodyStatus];

export const UpdateAppointmentBodyStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateAppointmentBody {
  status?: UpdateAppointmentBodyStatus;
  clientName?: string;
  clientPhone?: string;
  serviceId?: string;
  date?: string;
  time?: string;
  barberId?: string | null;
}

export interface AvailableSlots {
  date: string;
  slots: string[];
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  token: string;
}

export interface DashboardSummary {
  todayRevenue: number;
  monthRevenue: number;
  todayAppointments: number;
  monthAppointments: number;
  pendingAppointments: number;
  completedToday: number;
}

export interface RevenueChartEntry {
  date: string;
  revenue: number;
  appointments: number;
}

export interface ServicesChartEntry {
  serviceId: string;
  serviceName: string;
  count: number;
  revenue: number;
}

export type CreateRecurringAppointmentBodyPeriod =
  (typeof CreateRecurringAppointmentBodyPeriod)[keyof typeof CreateRecurringAppointmentBodyPeriod];

export const CreateRecurringAppointmentBodyPeriod = {
  this_month: "this_month",
  next_2_months: "next_2_months",
} as const;

export interface CreateRecurringAppointmentBody {
  clientName: string;
  clientPhone: string;
  serviceId: string;
  time: string;
  /** Day of week: 0=Sunday, 1=Monday, ... 6=Saturday */
  weekday: number;
  period: CreateRecurringAppointmentBodyPeriod;
  /** Reference date to determine current month (YYYY-MM-DD) */
  startDate: string;
  barberId?: string | null;
}

export interface RecurringAppointmentResult {
  groupId: string;
  created: Appointment[];
  /** Dates that were skipped due to conflicts */
  skipped: string[];
}

export interface ErrorResponse {
  error: string;
}

export type ListBlockedSlotsParams = {
  /**
   * Filter by date (YYYY-MM-DD)
   */
  date?: string;
};

export type ListAppointmentsParams = {
  /**
   * Filter by specific date (YYYY-MM-DD)
   */
  date?: string;
  /**
   * Filter by period
   */
  period?: ListAppointmentsPeriod;
  status?: ListAppointmentsStatus;
  /**
   * Filter by barber ID
   */
  barberId?: string;
};

export type ListAppointmentsPeriod =
  (typeof ListAppointmentsPeriod)[keyof typeof ListAppointmentsPeriod];

export const ListAppointmentsPeriod = {
  day: "day",
  week: "week",
  month: "month",
} as const;

export type ListAppointmentsStatus =
  (typeof ListAppointmentsStatus)[keyof typeof ListAppointmentsStatus];

export const ListAppointmentsStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type GetAvailableSlotsParams = {
  /**
   * Date in YYYY-MM-DD format
   */
  date: string;
  /**
   * Service ID to check duration
   */
  serviceId: string;
  /**
   * Filter conflicts for a specific barber
   */
  barberId?: string;
};

export type ExportAppointmentsParams = {
  /**
   * Period filter
   */
  period?: ExportAppointmentsPeriod;
  barberId?: string;
};

export type ExportAppointmentsPeriod =
  (typeof ExportAppointmentsPeriod)[keyof typeof ExportAppointmentsPeriod];

export const ExportAppointmentsPeriod = {
  day: "day",
  week: "week",
  month: "month",
  all: "all",
} as const;