# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gedilson Rai Barbershop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /barbers:
    get:
      operationId: listBarbers
      tags: [barbers]
      summary: List all barbers
      responses:
        "200":
          description: List of barbers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Barber"
    post:
      operationId: createBarber
      tags: [barbers]
      summary: Create a new barber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBarberBody"
      responses:
        "201":
          description: Barber created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"
  /barbers/{id}:
    get:
      operationId: getBarber
      tags: [barbers]
      summary: Get a single barber
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Barber details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBarber
      tags: [barbers]
      summary: Update a barber
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBarberBody"
      responses:
        "200":
          description: Updated barber
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments with optional filters
      parameters:
        - in: query
          name: date
          schema:
            type: string
          description: Filter by specific date (YYYY-MM-DD)
        - in: query
          name: period
          schema:
            type: string
            enum: [day, week, month]
          description: Filter by period
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, completed, cancelled]
        - in: query
          name: barberId
          schema:
            type: string
          description: Filter by barber ID
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "409":
          description: Time slot already booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /appointments/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [appointments]
      summary: Get available time slots for a date and service
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
          description: Date in YYYY-MM-DD format
        - in: query
          name: serviceId
          required: true
          schema:
            type: string
          description: Service ID to check duration
        - in: query
          name: barberId
          schema:
            type: string
          description: Filter conflicts for a specific barber
      responses:
        "200":
          description: List of available time slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableSlots"
...
  /appointments/export:
    get:
      operationId: exportAppointments
      tags: [appointments]
      summary: Export appointments as CSV
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [day, week, month, all]
          description: Period filter
        - in: query
          name: barberId
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /appointments/group/{groupId}:
    delete:
      operationId: deleteRecurringGroup
      tags: [appointments]
      summary: Delete all appointments in a recurrence group
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get a single appointment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment (status, etc.)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Barber:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        photo:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        birthDate:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        specialty:
          type: string
          nullable: true
        instagram:
          type: string
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
      required: [id, name, active, createdAt]
    CreateBarberBody:
      type: object
      properties:
        name:
          type: string
        photo:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        specialty:
          type: string
          nullable: true
      required: [name]
    UpdateBarberBody:
      type: object
      properties:
        name:
          type: string
        photo:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        birthDate:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        specialty:
          type: string
          nullable: true
        instagram:
          type: string
          nullable: true
        active:
          type: boolean
    Appointment:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
        serviceId:
          type: string
        serviceName:
          type: string
        servicePrice:
          type: number
        date:
          type: string
        time:
          type: string
        status:
          type: string
          enum: [pending, completed, cancelled]
        barberId:
          type: string
          nullable: true
        isRecurring:
          type: boolean
        recurrenceType:
          type: string
          nullable: true
        recurrenceGroupId:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, clientName, clientPhone, serviceId, serviceName, servicePrice, date, time, status, isRecurring, createdAt]
    CreateAppointmentBody:
      type: object
      properties:
        clientName:
          type: string
        clientPhone:
          type: string
        serviceId:
          type: string
        date:
          type: string
        time:
          type: string
        barberId:
          type: string
          nullable: true
      required: [clientName, clientPhone, serviceId, date, time]
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, completed, cancelled]
        clientName:
          type: string
        clientPhone:
          type: string
        serviceId:
          type: string
        date:
          type: string
        time:
          type: string
        barberId:
          type: string
          nullable: true
    AvailableSlots:
      type: object
      properties:
        date:
          type: string
        slots:
          type: array
          items:
            type: string
      required: [date, slots]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]